pyproject.toml
[tool.poetry]
name = "lyfe"
version = "0.1.0"
description = "bring your ideas to life"
authors = ["rr-hari-prakash <hari.prakash@rakuten.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
supabase = "^2.5.1"
python-dotenv = "^1.0.1"
pytest = "^8.2.2"
pytest-asyncio = "^0.23.7"
google-generativeai = "^0.7.1"
google-auth = "^2.32.0"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.137.0"
python-dateutil = "^2.9.0.post0"
pytz = "^2024.1"
requests = "^2.32.3"
notion-client = "^2.2.1"
aiohttp = "^3.9.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
